requirements.txt
azure-iot-device
azure-iot-hub
requests
asyncio
uvloop
aiohttp
aioredis
setuptools
redis